Testing before release
Dual wield implemented.

Chris:
"Known issue:
Right now, a weapon pickup will not automatically bring up a new weapon in the left hand if it is already equipped in the right.
For example. Player has shotgun equipped. A second shotgun does not come up if the player picks up a second shotgun. The player will need to cycle through the weapons back to the shotgun in order for the second shotgun to come up.

Instructions

1. The player has 2 playerweaponmanager components on it. There is a new public variable called Hand. The first weaponmanager needs to be set to RIGHT. The second one needs to be set to LEFT. They are only public to make referencing which one is which easier, If this is not set, or if this is changed, the whole thing will break.

2.  It should correct itself, but for ideal functionality, the left hand should not start with a weapon, or at least not a different one than is equipped in the right. This can be done by setting the size of the startingweapon array on the left weapon component to 0. "